requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
multidict==4.7.6
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-telegram-bot==12.8
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
typing-extensions==3.7.4.2
yarl==1.5.0
